--- sunet_virtual_env/pyproject.toml
@@ -1,23 +1,0 @@
-[tool.poetry]
-name = "sunet-virtual-env"
-version = "0.1.0"
-description = ""
-authors = ["Radu_Tolontan <radut@berkeley.edu>"]
-readme = "README.md"
-
-[tool.poetry.dependencies]
-python = "^3.11"
-matplotlib = "^3.9.3"
-numpy = "^2.1.3"
-scipy = "^1.14.1"
-pygame = "^2.6.1"
-pyaudio = "^0.2.14"
-sounddevice = "^0.5.1"
-construct = "^2.10.70"
-pyserial = "^3.5"
-crc16 = "^0.1.1"
-
-
-[build-system]
-requires = ["poetry-core"]
-build-backend = "poetry.core.masonry.api"
